Component for ProjectInfo
